requirements.txt
requests
beautifulsoup4
pandas
python-arango
pyyaml
schedule